admin: block access to certifications page when user has METIER role
