+++ admin/tests/unit/routes/authenticated/certifications/certification_test.js
@@ -3,28 +3,13 @@ import { setupTest } from 'ember-qunit';
 import sinon from 'sinon';
 import Service from '@ember/service';
 
-module('Unit | Route | authenticated/certifications/certification', function (hooks) {
+module('Unit | Route | authenticated/certifications', function (hooks) {
   setupTest(hooks);
-
-  test('#setupController', function (assert) {
-    // given
-    const certifications = { inputId: 5 };
-    const id = Symbol('id');
-    const route = this.owner.lookup('route:authenticated/certifications/certification');
-
-    // when
-    route.setupController(certifications, { id });
-
-    // then
-    // TODO: Fix this the next time the file is edited.
-    // eslint-disable-next-line qunit/no-assert-equal
-    assert.equal(certifications.inputId, id);
-  });
 
   module('#beforeModel', function () {
     test('it should check if current user is super admin, certif, or support', function (assert) {
       // given
-      const route = this.owner.lookup('route:authenticated/certifications/certification');
+      const route = this.owner.lookup('route:authenticated/certifications');
 
       const restrictAccessToStub = sinon.stub().returns();
       class AccessControlStub extends Service {
@@ -38,22 +23,5 @@ module('Unit | Route | authenticated/certifications/certification', function (ho
       // then
       assert.ok(restrictAccessToStub.calledWith(['isSuperAdmin', 'isCertif', 'isSupport'], 'authenticated'));
     });
-  });
-
-  test('#error', function (assert) {
-    // given
-    const route = this.owner.lookup('route:authenticated/certifications/certification');
-    const errorNotifierStub = {
-      notify: sinon.stub().resolves(),
-    };
-    route.errorNotifier = errorNotifierStub;
-    route.transitionTo = () => {};
-
-    // when
-    route.send('error');
-
-    // then
-    sinon.assert.called(errorNotifierStub.notify);
-    assert.ok(route);
   });
 });
